PBLmerged/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.0
email-validator>=2.0.0
python-multipart==0.0.12
